Add read-only access to the firmware version in conveyor::Printer
The firmware information can be accessed with
Printer::firmwareVersion(), which will return a small structure
containing three values: the major and minor firmware version and an
error code. The major and minor version are only valid if the error
code is kFirmwareVersionOK.

The FirmwareVersion::str() method returns a version string in the
format of "X.Y" or an error string.